Relocate BluetoothServiceInfo to helpers.service_info
Followup to #74418
